# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AccountDTO:
    properties:
      balance:
        description: em centavos
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  dto.CategoryBreakdownDTO:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      percentage:
        type: number
      total_amount:
        description: em centavos
        type: integer
      transaction_count:
        type: integer
    type: object
...
  dto.CreateAccountRequest:
    properties:
      currency:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - checking
        - savings
        - credit
        - investment
        type: string
    required:
    - currency
    - name
    - type
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.IncomeExpenseByPeriodResponse:
    properties:
      balance:
        description: em centavos
        type: integer
      end_date:
        type: string
      start_date:
        type: string
      total_expense:
        description: em centavos
        type: integer
      total_income:
        description: em centavos
        type: integer
    type: object
...
  dto.MonthlyTrendDTO:
    properties:
      expense:
        description: em centavos
        type: integer
      income:
        description: em centavos
        type: integer
      month:
        description: 'formato: "2006-01"'
        type: string
    type: object
...
  dto.UpdateAccountRequest:
    properties:
      currency:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - checking
        - savings
        - credit
        - investment
        type: string
    required:
    - currency
    - name
    - type
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API do sistema Organiza Aqui
  title: Organiza Aqui API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      description: Retorna todas as contas do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AccountDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar contas
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Cria uma nova conta financeira
      parameters:
      - description: Dados da conta
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AccountDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar conta
      tags:
      - accounts
  /api/v1/accounts/{id}:
    delete:
      description: Remove uma conta do sistema
      parameters:
      - description: ID da conta
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar conta
      tags:
      - accounts
    get:
      description: Retorna uma conta específica
      parameters:
      - description: ID da conta
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter conta
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma conta
      parameters:
      - description: ID da conta
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados da conta
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar conta
      tags:
      - accounts
  /api/v1/analytics/category-breakdown:
    get:
      description: Retorna o total gasto/recebido por categoria
      parameters:
      - description: Data inicial (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: Data final (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - description: Tipo (income ou expense)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryBreakdownDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Breakdown por Categoria
      tags:
      - analytics
  /api/v1/analytics/income-expense:
    get:
      description: Retorna o total de receitas e despesas em um período
      parameters:
      - description: Data inicial (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: Data final (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncomeExpenseByPeriodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Receitas e Despesas por Período
      tags:
      - analytics
  /api/v1/analytics/monthly-trend:
    get:
      description: Retorna receitas e despesas por mês
      parameters:
      - description: 'Número de meses (padrão: 6)'
        in: query
        name: months
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MonthlyTrendDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Tendência Mensal
      tags:
      - analytics
...
swagger: "2.0"